rename delete action
