+++ src/ducks/Watchlists/Actions/Delete/index.js
